Stage 9 - Integrate typeahead

+++ Vidly/Vidly/Dtos/NewRentalDto.cs
@@ -7,7 +7,7 @@ namespace Vidly.Dtos
 {
     public class NewRentalDto
     {
-        public int CustomerId { get; set; }
+        public byte CustomerId { get; set; }
         public List<int> MovieIds { get; set; }
     }
 }
